Revert "Backup Support for SQL Server in Azure VM"

+++ src/RecoveryServices/RecoveryServices.Backup.Providers/Providers/AzureSqlPsBackupProvider.cs
@@ -82,28 +82,18 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
             string containerUri = HelperUtils.GetContainerUri(keyValueDict, itemBase.Id);
             string protectedItemUri = HelperUtils.GetProtectedItemUri(keyValueDict, itemBase.Id);
 
-            throw new Exception(Resources.AzureSqlRetainDataException);
-        }
-
-        public RestAzureNS.AzureOperationResponse DisableProtectionWithDeleteData()
-        {
-            string vaultName = (string)ProviderData[VaultParams.VaultName];
-            string resourceGroupName = (string)ProviderData[VaultParams.ResourceGroupName];
-            bool deleteBackupData = (bool)ProviderData[ItemParams.DeleteBackupData];
-
-            ItemBase itemBase = (ItemBase)ProviderData[ItemParams.Item];
-            // do validations
-
-            ValidateAzureSQLDisableProtectionRequest(itemBase);
-
-            Dictionary<UriEnums, string> keyValueDict = HelperUtils.ParseUri(itemBase.Id);
-            string containerUri = HelperUtils.GetContainerUri(keyValueDict, itemBase.Id);
-            string protectedItemUri = HelperUtils.GetProtectedItemUri(keyValueDict, itemBase.Id);
-            return ServiceClientAdapter.DeleteProtectedItem(
+            if (deleteBackupData)
+            {
+                return ServiceClientAdapter.DeleteProtectedItem(
                     containerUri,
                     protectedItemUri,
                     vaultName: vaultName,
                     resourceGroupName: resourceGroupName);
+            }
+            else
+            {
+                throw new Exception(Resources.AzureSqlRetainDataException);
+            }
         }
 
         public RestAzureNS.AzureOperationResponse TriggerBackup()
@@ -495,16 +485,6 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
             return itemModels;
         }
 
-        public void RegisterContainer()
-        {
-            throw new NotImplementedException();
-        }
-
-        public List<PointInTimeBase> GetLogChains()
-        {
-            throw new NotImplementedException();
-        }
-
         #region private
         private void ValidateAzureSqlWorkloadType(CmdletModel.WorkloadType type)
         {
@@ -544,6 +524,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
                         typeof(CmdletModel.SimpleRetentionPolicy).ToString()));
             }
 
+            // call validation
             policy.Validate();
         }
